Add stone variant recipes and purpur from fruit to Aesthetics
This commit adds the much-needed andesite, diorite, and granite to the aesthetics factory. These 3 variants of stone did not generate in most biomes due to a config error, and an easy way to manufacture them is needed. The 3 recipes added stay as close to the ratios imposed by vanilla Minecraft as possible, but make one vital change: you can use quartz blocks made by hand OR by the aesthetics factory to make these stones, reducing the cost for these aesthetic blocks. It also introduces a 50% discount on quartz to make it favourable to vanilla crafting.

The 4th recipe added adds a way to make purpur the vanilla way: from cooked ('popped') chorus fruit. This follows the netherbrick crafting recipe cost (25% of vanilla cost).