Remove aggregate special stone recipes
The special stones, Andesite, Granite, and Diorite, are already abundant on the server. They are not available in all biomes, according to @CaptainKlutz due to a config error, however regardless of the cause this does not cause issues, and instead actually makes the game more interesting in the spirit of RealisticBiomes- people who don't live near mesa or savanna biomes will need to trade with those that do. Markets for special stone already exist in the two largest cities, Mount Augusta and Yoahtl, addressing this need organically, and adding this recipe hurts the people who invested time solving this shortage in-game. One year into the server being alive, we should not be adding config recipes for resources that are already possible to obtain in-game, and instead should only add recipes for those few that can't be.